refactor(w2.2): phase-split handle_delegate_task; mark W2.6 done
Move delegate_task choreography into comms_delegate.rs as ordered stages
(resolve → validate → compose → lease → create → send → track) while
preserving failure order and re-exporting the handler from comms.rs.
W2.6 resize contract already lives in render/resize.rs — document both
items complete in REFACTOR-PLAN and architecture maps.
